abstract facets builder, #https://github.com/TheScienceMuseum/collectionsonline/pull/140/files#r70432229
