+++ lib/transforms/search-results-to-jsonapi.js
@@ -4,6 +4,7 @@
 
 const QueryString = require('querystring');
 const TypeMapping = require('../type-mapping');
+const createFacets = require('../facets/create-facets');
 
 module.exports = (queryParams, results, config) => {
   results = results || {};
@@ -242,7 +243,7 @@ function createMeta (queryParams, results) {
   const total = (results.hits || {}).total || 0;
   const totalPages = Math.ceil(total / queryParams.pageSize);
   const countCategories = createCountCategories(results);
-  return { total_pages: totalPages, count: {type: countCategories}, filters: createFacets(queryParams.type, results) };
+  return { total_pages: totalPages, count: {type: countCategories}, filters: createFacets[queryParams.type](results) };
 }
 
 function createCountCategories (results) {
@@ -258,137 +259,6 @@ function createCountCategories (results) {
   result.objects += numbers.objects ? numbers.objects : 0;
   result.documents += numbers.documents ? numbers.documents : 0;
   return result;
-}
-
-/**
-* Create the facets object depending of the type of hte search (all, people, objects, documents)
-*/
-function createFacets (type, results) {
-  if (type === 'people') {
-    return createPeopleFacets(results);
-  }
-
-  if (type === 'documents') {
-    return createDocumentFacets(results);
-  }
-
-  if (type === 'objects') {
-    return createObjectFacets(results);
-  }
-
-  if (type === 'all') {
-    return createAllFacets(results);
-  }
-  return {};
-}
-
-function createAllFacets (results) {
-  var categoryValues = [];
-  results.aggregations.all.category.category_filters.buckets.forEach(value => {
-    categoryValues.push({ value: value.key, count: value.doc_count });
-  });
-
-  var makerValues = [];
-  results.aggregations.all.maker.maker_filters.buckets.forEach(value => {
-    makerValues.push({ value: value.key, count: value.doc_count });
-  });
-
-  var typeValues = [];
-  results.aggregations.all.type.type_filters.buckets.forEach(value => {
-    typeValues.push({ value: value.key, count: value.doc_count });
-  });
-
-  var placeValues = [];
-  results.aggregations.all.place.place_filters.buckets.forEach(value => {
-    placeValues.push({ value: value.key, count: value.doc_count });
-  });
-
-  var userValues = [];
-  results.aggregations.all.user.user_filters.buckets.forEach(value => {
-    userValues.push({ value: value.key, count: value.doc_count });
-  });
-
-  return {
-    category: categoryValues,
-    maker: makerValues,
-    type: typeValues,
-    place: placeValues,
-    user: userValues
-  };
-}
-
-function createPeopleFacets (results) {
-  var occupationValues = [];
-  results.aggregations.people.occupation.occupation_filters.buckets.forEach(value => {
-    occupationValues.push({ value: value.key, count: value.doc_count });
-  });
-
-  var placeBorn = [];
-  results.aggregations.people.place_born.place_born_filters.buckets.forEach(value => {
-    placeBorn.push({ value: value.key, count: value.doc_count });
-  });
-
-  var organisation = [];
-  results.aggregations.people.organisation.organisations_filters.buckets.forEach(value => {
-    organisation.push({ value: value.key, count: value.doc_count });
-  });
-
-  return {
-    occupation: occupationValues,
-    place_born: placeBorn,
-    organisation: organisation
-  };
-}
-
-function createObjectFacets (results) {
-  var categoryValues = [];
-  results.aggregations.objects.category.category_filters.buckets.forEach(value => {
-    categoryValues.push({ value: value.key, count: value.doc_count });
-  });
-
-  var makerValues = [];
-  results.aggregations.objects.maker.maker_filters.buckets.forEach(value => {
-    makerValues.push({ value: value.key, count: value.doc_count });
-  });
-
-  var typeValues = [];
-  results.aggregations.objects.type.type_filters.buckets.forEach(value => {
-    typeValues.push({ value: value.key, count: value.doc_count });
-  });
-
-  var placeValues = [];
-  results.aggregations.objects.place.place_filters.buckets.forEach(value => {
-    placeValues.push({ value: value.key, count: value.doc_count });
-  });
-
-  var userValues = [];
-  results.aggregations.objects.user.user_filters.buckets.forEach(value => {
-    userValues.push({ value: value.key, count: value.doc_count });
-  });
-
-  return {
-    category: categoryValues,
-    maker: makerValues,
-    type: typeValues,
-    place: placeValues,
-    user: userValues
-  };
-}
-
-function createDocumentFacets (results) {
-  var archiveValues = [];
-  results.aggregations.documents.archive.archive_filters.buckets.forEach(value => {
-    archiveValues.push({ value: value.key, count: value.doc_count });
-  });
-
-  var organisationValues = [];
-  results.aggregations.documents.organisation.organisation_filters.buckets.forEach(value => {
-    organisationValues.push({ value: value.key, count: value.doc_count });
-  });
-  return {
-    archive: archiveValues,
-    organisation: organisationValues
-  };
 }
 
 function xtend (obj1, obj2) {
